Fix statcord
There was a subtle scoping issue here. The line `statcord.on` expects 
`statcord` to refer to the globally defined variable. But there are a 
lot of scopes to check before global. It was instead finding the 
module-level function `statcord`. Was not a problem before because the 
function was (essentially) called `module.exports.statcord`.